api/requirements.txt
Flask
flask_sqlalchemy
flask_cors
requests
python-dotenv
psycopg2-binary